reactdjango/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-multiupload==0.6.1
et-xmlfile==1.1.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
PyPDF2==3.0.1
PyPDF4==1.27.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
